5주차 미션: 컴포넌트 분리
