+++ week3/kimminji/src/components/Setting/TimeSettingGroup.module.css
@@ -1,0 +1,31 @@
+.settingGroup {
+    background-color: white;
+    padding: 2rem;
+    border-radius: 1.5rem;
+    box-shadow: 0 8px 20px rgba(0, 0, 0, 0.08);
+    flex-shrink: 0;
+  }
+  
+  .settingGroup > label {
+    background-color: white;
+  }
+  
+  .dropdownGroup {
+    display: flex;
+    gap: 10px;
+    background-color: white;
+    margin-bottom: 1.5rem;
+  }
+  
+  input[type="number"],
+  select {
+    width: 100%;
+    padding: 0.6rem;
+    margin-top: 0.3rem;
+    margin-bottom: 1rem;
+    border-radius: 0.5rem;
+    border: 1px solid #ccc;
+    background-color: white;
+    font-size: 1rem;
+  }
+
