+++ week3/kimminji/src/components/Setting/VideoSearch.module.css
@@ -1,0 +1,74 @@
+.videoSection {
+    background-color: white;
+    padding: 2rem;
+    border-radius: 1.5rem;
+    box-shadow: 0 8px 20px rgba(0, 0, 0, 0.08);
+    flex-shrink: 0;
+  }
+  
+  .videoFrame {
+    background-color: white;
+    max-height: 400px;
+  }
+  
+  .videoFrame > h3 {
+    background-color: white;
+  }
+  
+  input[type="text"] {
+    width: 95%;
+    padding: 0.6rem;
+    margin-top: 0.3rem;
+    margin-bottom: 1rem;
+    border-radius: 0.5rem;
+    border: 1px solid #ccc;
+    background-color: white;
+    font-size: 1rem;
+  }
+  
+  .videoList {
+    display: flex;
+    flex-direction: column;
+    gap: 1rem;
+    height: 270px;
+    overflow-y: auto;
+    background-color: white;
+  }
+  
+  .videoItem {
+    display: flex;
+    align-items: flex-start;
+    gap: 1rem;
+    padding: 1rem;
+    border-radius: 1rem;
+    border: 1px solid #ccc;
+    width: 100%;
+    background-color: white;
+    text-align: left;
+    cursor: pointer;
+  }
+  
+  .videoItem:hover {
+    background-color: #e0e0e0;
+  }
+  
+  .selected {
+    background-color: #d1e7ff;
+  }
+  
+  .videoList > div {
+    background-color: white;
+  }
+  
+  .videoTitle {
+    font-weight: bold;
+    margin-bottom: 0.2rem;
+    background-color: transparent;
+  }
+  
+  .videoDesc {
+    font-size: 0.9rem;
+    color: #555;
+    background-color: transparent;
+  }
+
