+++ week3/kimminji/src/pages/Setting.module.css
@@ -2,15 +2,6 @@ body {
   background-color: #f9f9f9;
 
 }
-
-.header {
-  width: 100%;
-  text-align: center;
-  padding: 2rem 1rem;
-  margin-bottom: 2rem;
-  border-radius: 1rem;
-}
-
 .container {
   display: flex;
   width: 100%;
@@ -21,107 +12,19 @@ body {
   align-items: flex-start;
 }
 
+.header {
+  width: 100%;
+  text-align: center;
+}
+
 .innercontainer {
   display: flex;
   flex-direction: row;
   gap: 2rem;
   width: 100%;
+  height: 450px;
   justify-content: center;
-}
-
-.settingGroup, .videoSection {
-  background-color: white;
-  padding: 2rem;
-  border-radius: 1.5rem;
-  box-shadow: 0 8px 20px rgba(0, 0, 0, 0.08);
-  flex-shrink: 0;
-}
-.settingGroup > label {
-  background-color: white;
-}
-
-.dropdownGroup {
-  display: flex;
-  gap: 10px;
-  background-color: white;
-  margin-bottom: 1.5rem;
-}
-
-input[type="number"], select, input[type="text"] {
-  width: 100%;
-  padding: 0.6rem;
-  margin-top: 0.3rem;
-  margin-bottom: 1rem;
-  border-radius: 0.5rem;
-  border: 1px solid #ccc;
-  background-color: white;
-  font-size: 1rem;
-}
-
-input[type="text"] {
-  width: 95%;
-  padding: 0.6rem;
-  margin-top: 0.3rem;
-  margin-bottom: 1rem;
-  border-radius: 0.5rem;
-  border: 1px solid #ccc;
-  background-color: white;
-  font-size: 1rem;
-}
-
-.videoFrame {
-  background-color: white;
-}
-
-.videoFrame > h3 {
-  background-color: white;
-}
-
-.videoList {
-  display: flex;
-  flex-direction: column;
-  gap: 1rem;
-  max-height: 400px;
-  overflow-y: auto;
-  background-color: white;
-}
-
-.videoItem {
-  display: flex;
-  align-items: flex-start;
-  gap: 1rem;
-  padding: 1rem;
-  border-radius: 1rem;
-  border: 1px solid #ccc;
-  width: 100%;
-  background-color: white;
-  text-align: left;
-  cursor: pointer;
-}
-
-.videoItem:hover {
-  background-color: #e0e0e0;
-}
-
-.selected {
-  background-color: #d1e7ff;
-}
-
-.videoList > div {
-  background-color: white;
-}
-
-.videoTitle {
-  font-weight: bold;
-  margin-bottom: 0.2rem;
-  background-color: transparent;
-}
-
-.videoDesc {
-  font-size: 0.9rem;
-  color: #555;
-  background-color: transparent;
-
+  /* border: 1px solid #ddd; */
 }
 
 .buttonWrapper {
